==> 1_quick_exploration.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

pd.set_option('display.max_columns', 500)
pd.set_option('display.max_rows', 500)

data_down = pd.read_csv('Data_down.csv', sep='\s+')  #your oath here 
data_up = pd.read_csv('Data_up.csv', sep='\s+')    #your oath here 

data_down.head()

data_up.head()

data_up.columns

data_down.columns

data_down.dtypes

data_up.dtypes

data_down.describe()

data_down.isnull().values.any()

data_up.isnull().values.any()

locations_list = list(data_down['location'].unique())
def group_count(df, locations, groupby_1, groupby_2,groupby_3 ):
    
    describtion = pd.DataFrame(columns =['location', 'measurement_count', 'pci_count', 'ci_count'])
    for loc in locations:
        location = df[df['location'] == loc]
        
        describtion= describtion.append({
            'location': loc,
            'measurement_count': location.groupby(groupby_1).count().shape[0],
            'pci_count': location.groupby(groupby_2).count().shape[0],
            'ci_count': location.groupby(groupby_3).count().shape[0]
        }, 
            ignore_index=True)
        
    return describtion

down_unique_describtion = group_count(data_down,locations_list, 'measurement', 'pci', 'ci')
down_unique_describtion 

upload_unique_describtion = group_count(data_up,locations_list, 'measurement', 'pci', 'ci')
upload_unique_describtion

#  visualizations

def plot_2_hist(df1, df2, x1, x2, title1 ="", title2=""):

    _, axes = plt.subplots(1, 2, figsize=(16, 8))

    ax1 = sns.histplot(data=df1, x=x1,
                    kde=True, ax=axes[0])
    ax1.set_title(title1, loc='left')

    ax2 = sns.histplot(data=df2, x=x2,
                    kde=True, ax=axes[1])
    ax2.set_title(title2, loc='left')


    plt.show()
    
def plot_2_boxplot(x1, x2, title1 ="", title2=""):

    _, axes = plt.subplots(1, 2, figsize=(16, 8))

    ax1 = sns.boxplot(x=x1, ax=axes[0])
    ax1.set_title(title1, loc='left')

    ax2 = sns.boxplot(x=x2, ax=axes[1])
    ax2.set_title(title2, loc='left')


    plt.show()

plot_2_hist(data_down, data_up,data_down["throughput"], data_up["throughput"],
            title1 = "Download Throughput", title2 = "Upload Throughput")

plot_2_hist(data_down, data_up,data_down["speed"], data_up["speed"],
            title1 = "Spead Download-Data", title2 = "Spead Upload-Data")

plot_2_hist(data_down, data_up,data_down["acc"], data_up["acc"],
            title1 = "Acceleration Download-Data", title2 = "Acceleration Upload-Data")

plot_2_boxplot(data_down["throughput"], data_up["throughput"],
              title1 = "Download Throughout", title2="Upload throughput")

plot_2_boxplot(data_down["speed"], data_up["speed"],
              title1 = "Spead Download-Data", title2="Spead Upload-Data")
