==> 4_model_trials.ipynb <==
#!pip install xgboost

import pandas as pd
import numpy as np

from my_defs import*
from sklearn.model_selection import train_test_split

seed = 42
pd.set_option('display.max_columns', 500)

data_down = pd.read_csv('data_down_full.csv')
data_up = pd.read_csv('data_up_full.csv')

drop_cols_u  = [ 'connected', 'date', 'year', 'rawTimesamp',
             'month', 'id', 'throughput_var', 'ci',
           'tavg', 'tmax', 'tmin', 'wdir' ,'pres', 'tsun','dir',
                'campus', 'highway', 'urban', 'suburban',
           'Barop', 'Brünninghausen', 'Eichlinghofen', 'Groß-Barop', 'Hombruch', 'Innenstadt Nord',
           'KGV Ruhrwaldstraße', 'Kirchhörde', 'Klinikviertel', 'Kruckel', 'Löttringhausen',
           'Lücklemberg', 'Mitte' , 'Persebeck', 'Renninghausen', 'Salingen', 'Syburg', 'Wellinghofen',
            'Wichlinghofen']

drop_cols_d  = ['connected', 'date', 'year','rawTimesamp',
             'month', 'id','throughput_var', 'ci',
           'tavg', 'tmax', 'tmin', 'wdir' ,'pres', 'tsun','campus', 'highway', 'urban', 'suburban', 'dir',
           'Barop', 'Brünninghausen', 'Eichlinghofen', 'Groß-Barop', 'Hombruch', 
           'KGV Ruhrwaldstraße', 'Kirchhörde', 'Klinikviertel', 'Kruckel', 'Löttringhausen',
           'Lücklemberg', 'Mitte' , 'Persebeck', 'Renninghausen', 'Salingen', 'Syburg', 'Wellinghofen',
            'Wichlinghofen']
            
data_down = data_down.drop(drop_cols_d, axis = 1)
data_up = data_up.drop(drop_cols_u, axis = 1)



x_train_d,x_test_d,y_train_d,y_test_d=train_test_split(data_down.drop(['throughput'], axis = 1)
                                               ,data_down["throughput"], test_size=0.2,
                                                      random_state = seed,
                                                      shuffle = False)

x_train_u,x_test_u,y_train_u,y_test_u=train_test_split(data_up.drop(['throughput'], axis = 1)
                                               ,data_up["throughput"], test_size=0.2,
                                                      random_state= seed,
                                                      shuffle = False)

# XGBoost

regressor_rmse(xgb.XGBRegressor(random_state = seed),x_train_d,   # function in my_defs
               x_test_d, y_train_d, y_test_d, model_name = "XGBoost Download")

regressor_rmse(xgb.XGBRegressor(random_state = seed),x_train_u,   # function in my_defs
               x_test_u, y_train_u, y_test_u, model_name = "XGBoost Upload")

## Random Forest

regressor_rmse(RandomForestRegressor(random_state = seed) ,x_train_d,  
               x_test_d, y_train_d, y_test_d, model_name = "Random Forrest Download")

regressor_rmse(RandomForestRegressor(random_state = seed) ,x_train_u, 
               x_test_u, y_train_u, y_test_u, model_name = "Random Forrest Upload")

# Decision Tree

regressor_rmse(DecisionTreeRegressor(random_state = seed) ,x_train_d, 
               x_test_d, y_train_d, y_test_d, model_name = "Decision Tree Download")

regressor_rmse(DecisionTreeRegressor(random_state = seed) ,x_train_u, 
               x_test_u, y_train_u, y_test_u, model_name = "Decision Tree Upload")


# Lasso Reg 

regressor_rmse(Lasso(random_state = seed) ,x_train_d, 
               x_test_d, y_train_d, y_test_d, model_name = "Lasso Regression Download")

regressor_rmse(DecisionTreeRegressor(random_state = seed) ,x_train_u, 
               x_test_u, y_train_u, y_test_u, model_name = "Decision Tree Upload")

# SVR


regressor_rmse(SVR() ,x_train_d,  #SVR has no random_state parameter
               x_test_d, y_train_d, y_test_d, model_name = "Support Vector Regression Download")

regressor_rmse(SVR() ,x_train_u, #SVR has no random_state parameter
               x_test_u, y_train_u, y_test_u, model_name = "Support Vector Regression Upload")

## Linear Reg

# Download Model
regressor_rmse(LinearRegression() ,x_train_d, #SVR has no random_state parameter
               x_test_d, y_train_d, y_test_d, model_name = "Linear Regression Download")

# Download Model
regressor_rmse(LinearRegression() ,x_train_u, 
               x_test_u, y_train_u, y_test_u, model_name = "Linear Regression Upload")
